Extract price list version utility in common class
